AdminApp/serialiazer.py
from rest_framework import serializers
from .models import UserProfile, Student, Teacher

class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Student
        fields = ['id', 'user', 'reg_no', 'name', 'batch', 'email', 'course', 'contace']


class TeacherSerializer(serializers.ModelSerializer):
    class Meta:
        model = Teacher
        fields = ['id', 'user', 'reg_no', 'name', 'contact', 'email']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ['id', 'username', 'password']