requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
cssselect2==0.7.0
daphne==3.0.2
Django==4.2.5
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.42.1
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mysqlclient==2.2.0
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.7.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.5
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
weasyprint==59.0
webencodings==0.5.1
zope.interface==6.1
zopfli==0.2.3
